changed return type form int to unsigned long in rt_bits_reset, 64 bits safer

+++ base/ipc/bits/bits.c
@@ -194,7 +194,7 @@ RTAI_SYSCALL_MODE unsigned long rt_get_bits(BITS *bits)
 	return bits->mask;
 }
 
-RTAI_SYSCALL_MODE int rt_bits_reset(BITS *bits, unsigned long mask)
+RTAI_SYSCALL_MODE unsigned long rt_bits_reset(BITS *bits, unsigned long mask)
 {
 	unsigned long flags, schedmap, oldmask;
 	RT_TASK *task;
